Separate sessions commands into separate files
